feat: Remove use-client page, change system show-modal, resolve implementation traductions from topNav
