+++ src/app/[locale]/page.js
@@ -1,5 +1,4 @@
-"use client";
-import React, { useState, useEffect } from "react";
+import React from "react";
 import TopNavHero from "@/app/components/topNav";
 import WhatWeDo from "../components/whatWeDo";
 import ResponsibleCompany from "../components/responsibleCompany";
@@ -13,21 +12,15 @@ import OurCommit from "../components/ourCommit";
 import OurCommitWomen from "../components/ourCommitWomen";
 import TopProduct from "../components/topProduct";
 import NewProduct from "../components/newProduct";
-// import MapW from "../components/map";
 import { ProductCategories } from "../components/productCategories";
 import { WorkerTestimonials } from "../components/workerTestimonials";
-import ContactForm from "../components/modal/ContactForm";
-export default function Home() {
-  const [showModalContact, setShowModalContact] = useState(false);
-  const toggleModalContact = () => {
-    setShowModalContact((prev) => !prev);
-  };
-
+export default async function Home({ params }) {
+  const { locale } = await params;
   return (
     <div
       className={`bg-white/20 w-full max-w-[1920px] mx-auto items-center justify-items-center min-h-screen`}
     >
-      <TopNavHero toggleModalContact={toggleModalContact} />
+      <TopNavHero locale={locale} />
       <WhatWeDo />
       <ResponsibleCompany />
       <Experience />
@@ -44,9 +37,6 @@ export default function Home() {
       <OurCommit />
       <OurCommitWomen />
       <WorkerTestimonials />
-      {showModalContact && (
-        <ContactForm toggleModalContact={toggleModalContact} />
-      )}
     </div>
   );
 }
