+++ src/app/components/topNav.js
@@ -2,34 +2,24 @@
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 import { IoMdMenu } from "react-icons/io";
-import { useTranslations } from "next-intl"; // Importar useTranslations
+import { useTranslations } from "next-intl";
 import Link from "next/link";
 import { NavLinkMultiple } from "./navLinkMultiple";
+import ContactForm from "./modal/ContactForm";
 
-// const slides = [
-//   { type: "video", src: "/hero/video.mp4" }, // Asegúrate de que la ruta sea absoluta desde public
-//   { type: "image", src: "/hero/image2.jpg" }, // Asegúrate de que la ruta sea absoluta desde public
-// ];
 const CANT_LINKS_PRODUCTS = 4;
-export default function TopNavHero({ toggleModalContact }) {
-  // const [current, setCurrent] = useState(0);
+export default function TopNavHero({ locale }) {
   const [menuOpen, setMenuOpen] = useState(false);
-  const { locale, push } = useRouter();
-  const t = useTranslations("topNav"); // Inicializar useTranslations
-
-  // const nextSlide = () => setCurrent((prev) => (prev + 1) % slides.length);
-  // const prevSlide = () => setCurrent((prev) => (prev - 1 + slides.length) % slides.length);
-
-  const changeLanguage = (lang) => {
-    push("/", { locale: lang });
+  const [showModalContact, setShowModalContact] = useState(false);
+  const toggleModalContact = () => {
+    setShowModalContact((prev) => !prev);
   };
-
-  /*   link :
-    que hacemos: wedo 
-    nuestra calidad: ourquality
-    nuestro compromiso: ourcommit
-    blog: blog
-    */
+  const { push, refresh } = useRouter();
+  const t = useTranslations("topNav"); // Inicializar useTranslations
+  const changeLanguage = (lang) => {
+    push(lang);
+    refresh();
+  };
 
   return (
     <div className="w-full">
@@ -44,7 +34,6 @@ export default function TopNavHero({ toggleModalContact }) {
               <Link href="#" className="hover:underline hover:font-bold">{t("whatWeDo")}</Link> {/* Utilizando la traducción */}
             </li>
             <li>
-              {/* <Link href="#" className="hover:underline hover:font-bold">{t("products.title")}</Link> Utilizando la traducción */}
               <NavLinkMultiple title={t('products.title')} options={new Array(CANT_LINKS_PRODUCTS).fill(0).map((_, i) => {
                 return {
                   id: t(`products.options.${i}.id`),
@@ -80,54 +69,9 @@ export default function TopNavHero({ toggleModalContact }) {
           </button>
         </div>
       </nav>
-
-      {/* <div className="relative w-full h-screen overflow-hidden">
-        {slides.map((slide, index) => (
-          <motion.div
-            key={index}
-            className={`absolute w-full h-full transition-opacity duration-500 ${
-              index === current ? "opacity-100" : "opacity-0"
-            }`}
-          >
-            {slide.type === "image" ? (
-              <img
-                src={slide.src} // Rutas absolutas desde public
-                alt="Slide"
-                className="w-full h-full object-cover"
-              />
-            ) : (
-              <video autoPlay loop muted className="w-full h-full object-cover">
-                <source src={slide.src} type="video/mp4" />
-              </video>
-            )}
-          </motion.div>
-        ))}
-
-        <button
-          onClick={prevSlide}
-          className="absolute left-4 top-1/2 transform -translate-y-1/2 bg-black/50 p-2 rounded-full text-white"
-        >
-          <FaChevronLeft />
-        </button>
-        <button
-          onClick={nextSlide}
-          className="absolute right-4 top-1/2 transform -translate-y-1/2 bg-black/50 p-2 rounded-full text-white"
-        >
-          <FaChevronRight />
-        </button>
-
-        <div className="absolute bottom-4 left-1/2 transform -translate-x-1/2 flex gap-2">
-          {slides.map((_, index) => (
-            <button
-              key={index}
-              className={`w-3 h-3 rounded-full ${
-                index === current ? "bg-white" : "bg-gray-400"
-              }`}
-              onClick={() => setCurrent(index)}
-            />
-          ))}
-        </div>
-      </div> */}
+      {showModalContact && (
+        <ContactForm toggleModalContact={toggleModalContact} />
+      )}
     </div>
   );
 }
